+++ examples/settings/settings_history.py
@@ -2,16 +2,11 @@ from sedai import settings_history, settings
 from datetime import datetime, timedelta
 from typing import List
 
+from sedai.settings import SettingsConfigMode
+
 resource_id = "resourceid"
 group_id = "groupid"
 account_id = "accountid"
-
-
-def swap_mode(mode):
-    if mode == 'MANUAL':
-        return 'AUTO'
-    return 'MANUAL'
-
 
 def print_events(events: List[settings_history.SettingsChangeEvent]):
     print(f"{len(events)} event(s) received...")
@@ -26,37 +21,33 @@ def print_events(events: List[settings_history.SettingsChangeEvent]):
 
 # Modify resource settings
 resource_settings = settings.get_resource_settings(resource_id)
-resource_settings.availabilityMode = swap_mode(resource_settings.availabilityMode)
+resource_settings.availabilityMode = SettingsConfigMode.DATA_PILOT
 resource_settings = settings.update_resource_settings(resource_id, resource_settings)
 
-starttime = datetime.now() - timedelta(seconds=1)
+start_time = datetime.now() - timedelta(seconds=1)
 endtime = datetime.now()
-events = settings_history.get_resource_settings_history(resource_id, starttime, endtime)
+events = settings_history.get_resource_settings_history(resource_id, start_time, endtime)
 print("Resource settings events:")
 print_events(events)
 
 # Modify group settings
 group_settings = settings.get_group_settings(group_id)
-group_settings.app_settings.availabilityMode = swap_mode(
-    group_settings.app_settings.availabilityMode
-)
+group_settings.app_settings.availabilityMode = SettingsConfigMode.DATA_PILOT
 group_settings = settings.update_group_settings(group_id, group_settings)
 
-starttime = datetime.now() - timedelta(seconds=1)
+start_time = datetime.now() - timedelta(seconds=1)
 endtime = datetime.now()
-events = settings_history.get_group_settings_history(group_id, starttime, endtime)
+events = settings_history.get_group_settings_history(group_id, start_time, endtime)
 print("\nGroup settings events:")
 print_events(events)
 
 # Modify account settings
 account_settings = settings.get_account_settings(account_id)
-account_settings.app_settings.availabilityMode = swap_mode(
-    account_settings.app_settings.availabilityMode
-)
+account_settings.app_settings.availabilityMode = SettingsConfigMode.DATA_PILOT
 account_settings = settings.update_account_settings(account_id, account_settings)
 
-starttime = datetime.now() - timedelta(seconds=1)
+start_time = datetime.now() - timedelta(seconds=1)
 endtime = datetime.now()
-events = settings_history.get_account_settings_history(account_id, starttime, endtime)
+events = settings_history.get_account_settings_history(account_id, start_time, endtime)
 print("\nAccount settings events:")
 print_events(events)